fix(FIJI-5563): [[Edge Private mode] App cannot be opened due to 'IndexedDB' is not defined] delete database by daoManager

+++ packages/foundation/src/db/DBManager.ts
@@ -30,7 +30,9 @@ class DBManager {
   }
 
   async deleteDatabase(): Promise<void> {
-    await this.db.delete();
+    if (this.db) {
+      await this.db.delete();
+    }
   }
 
   isDatabaseOpen(): boolean {
